add extra explanation add extra syncronisation on the producer 	modified:   README.md 	modified:   receiver/virtuoso/receiver_cont_sync.sh 	modified:   receiver/virtuoso/receiver_only_sync.sh

+++ receiver/virtuoso/receiver_cont_sync.sh
@@ -5,20 +5,30 @@ INSTALLDIR=ENV_INSTALL_DIR/receiver/virtuoso
 
 if [ -f ENV_WORK_DIR/sync ] ;
 then
-    echo "start" ;
-    rm -f ENV_WORK_DIR/sync
-    echo "upload most recent data" ;
-    $INSTALLDIR/execute-isql.sh $INSTALLDIR/clean_upload.sql
-    echo "dump all data" ;
-    rm -rf ENV_VIRTUOSO_DUMP_DIR
-    $INSTALLDIR/execute-isql.sh $INSTALLDIR/dump_dmz.sql
-    sleep 10m
+    if [ -f ENV_WORK_DIR/receiving ] ; 
+    then
+	# there is already a receiving operation started. Wait until it is finished to start another one.
+	echo "." 
+    else
+	    echo "start" ;
+	    rm -f ENV_WORK_DIR/sync
+            cp ENV_INSTALL_DIR/receiver/receiver ENV_WORK_DIR/receiving
 
-    $INSTALLDIR/transferdumps.sh
+	    echo "upload most recent data" ;
+	    $INSTALLDIR/execute-isql.sh $INSTALLDIR/clean_upload.sql
+	    echo "dump all data" ;
+	    rm -rf ENV_VIRTUOSO_DUMP_DIR
+	    $INSTALLDIR/execute-isql.sh $INSTALLDIR/dump_dmz.sql
+	    # the sleep has to be checked.
+            # It can only be checked via implementing a poll on a virtuoso loading table 
+	    sleep 10m
 
-    scp ENV_INSTALL_DIR/producer/sync ENV_TARGET:ENV_TARGET_DIR
-    echo "done" ;
+	    $INSTALLDIR/transferdumps.sh
 
+	    scp ENV_INSTALL_DIR/producer/sync ENV_TARGET:ENV_TARGET_DIR
+	    echo "done" ;
+	    rm -f ENV_WORK_DIR/receiving
+    fi
 else 
     echo "." ;
 fi
